pidstat: Add PF_NO_SETAFFINITY const
For the stat flag that means that userland is not allowed to meddle with
cpus_allowed, i.e. with the thread's smp affinity, via
sched_setaffinity.

This indeed has the same value as PF_THREAD_BOUND, see the kernel
sources for an explanation, but basically was in this cset:

commit 14a40ffccd6163bbcd1d6f32b28a88ffe6149fc6
Author: Tejun Heo <tj@kernel.org>
Date:   Tue Mar 19 13:45:20 2013 -0700

    sched: replace PF_THREAD_BOUND with PF_NO_SETAFFINITY

    PF_THREAD_BOUND was originally used to mark kernel threads which
    were bound to a specific CPU using kthread_bind() and a task with
    the flag set allows cpus_allowed modifications only to itself.
    Workqueue is currently abusing it to prevent userland from meddling
    with cpus_allowed of workqueue workers.

 ---------------------------------------------------------------

So add the new const but keep the old one, we may have some tool out
there using it.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>