python-linux-procfs: bitmasklist_test.py Add the first unit test
This adds the first unit test to python-linux-procfs

It adds the bitmasklist_test for testing the function bitmasklist, which
is called by parse_affinity. Unlike parse_affinity, the number of cpus
is not detected, so the unit test can simulate different numbers of
cpus.

This is an expanded form of a reproducer from Jozef Bacik from
Bugzilla 1365902

If we add more unit tests in the future (we should), then we can
consider some kind of python package scheme, but for now this can live
in the base directory.

Signed-off-by: John Kacur <jkacur@redhat.com>
Signed-off-by: Jiri Kastner <jkastner@redhat.com>