python-linux-procfs: Propagate error to user if a pid is completed
If a pid is completed and disappears a FileNotFoundError will occur
because /proc/pid/stat will disappear too.

It is not possible to check for the file first because it could still
disappear between the time of the check and the time of use.

Propagate this error to the user.
The user should handle this with a try, except clause and ignore it if
an exception occurs.

Reviewed-by: Leah Leshchinsky <lleshchi@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>