python-linux-procfs: Fix UnicodeDecodeError
Commit 7570fc0d6082 meant to solve the UnicodeDecodeError

Instead it actually increased the problem by reading lines as bytes
and decoding them.

The original problem is hard to trigger and doesn't trigger consistently
with reproducers. In addition there seems to be a difference in how this
is handled between python-3.6 to python-3.9

For now, we should return the code to reading as utf-8 (the default)
since that handles more cases than the current code.

We can catch the UnicodeDecodeError and ignore it for now. It is not
ideal because we are not handling some pids that trigger the error.

This patch also includes a fix for a FileNotFoundError which can occur if
a pid exits and disappears before we try to read it in the /proc file
system.

Signed-off-by: John Kacur <jkacur@redhat.com>